requirements.txt
asgiref==3.8.1
click==8.1.7
click-log==0.4.0
cloudinary==1.41.0
cloudinary-cli==1.10.3
crispy-bootstrap4==2024.1
dj-database-url==0.5.0
Django==4.2.16
django-crispy-forms==2.3
django-js-asset==2.2.0
django-mptt==0.16.0
django-summernote==0.8.20.0
docstring_parser==0.16
psycopg2==2.9.9
sqlparse==0.5.1
